+++ CHANGELOG.md
@@ -11,7 +11,9 @@ This release updates the Git source path to this Terraform module since the GitL
   * **Breaking Change:** Update `README` example usage `module source` to new path.
   * Update `CHANGELOG` links to issues to use new path
 * [#12](https://gitlab.com/gitlab-com/demo-systems/terraform-modules/gcp/gce/gcp-compute-instance-tf-module/-/issues/12) - Update example usage GCP machine type from `n1-standard-2` to `e2-standard-2`.
+* [#14](https://gitlab.com/gitlab-com/demo-systems/terraform-modules/gcp/gce/gcp-compute-instance-tf-module/-/issues/14) - Update README to add clarity on Terraform GCP provider authentication with environment variables and key files
 * [#15](https://gitlab.com/gitlab-com/demo-systems/terraform-modules/gcp/gce/gcp-compute-instance-tf-module/-/issues/15) - Fix README with example usage label for `gl_env_name` to use alphadash syntax
+* [#16](https://gitlab.com/gitlab-com/demo-systems/terraform-modules/gcp/gce/gcp-compute-instance-tf-module/-/issues/16) - Fix example usage outputs that is not a valid attribute of returned resource object
 
 ## 0.3.0